requirements.txt
numpy
pandas
tqdm
opencv-python
Pillow
pyyaml
scipy
torch
torchvision
torchaudio
scikit-image
albumentations
matplotlib
scikit-learn
pytorch-grad-cam
jupyter
ipykernel

